+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_newman_tobias.py
@@ -181,7 +181,7 @@ class TestNewmanTobiasWithSEI(unittest.TestCase):
         model.check_well_posedness()
 
 
-class TestNewmanTobiasWithCrack(unittest.TestCase):
+class TestNewmanTobiasWithMechanics(unittest.TestCase):
     def test_well_posed_negative_cracking(self):
         options = {"particle mechanics": ("swelling and cracking", "none")}
         model = pybamm.lithium_ion.NewmanTobias(options)